requirements.txt
functions-framework==3.5.0
python-telegram-bot==21.1.1
replicate==0.28.0
langchain==0.2.7
langchain_community==0.2.7
reportlab==4.2.2
google-api-python-client==2.84.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
pandas==2.0.0
testresources==2.0.1
openai==1.35.13
